Role-Permission-Annoucement-mingration departement, study program

+++ app/Http/Controllers/DepartmentController.php
@@ -2,35 +2,38 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Departement;
+use App\Models\Department;
 use Illuminate\Http\Request;
 use Yajra\DataTables\Facades\DataTables;
 
-class DepartementController extends Controller
+class DepartmentController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+    public function __construct() {
+        $this->middleware('auth');
+    }
     public function index(Request $request)
     {
+        $this->authorize('setting/manage_data/department.read');
         if ($request->isMethod('POST')) { 
             $this->validate($request, [ 
                 'name_dept' => 'required',
             ]);
-            $new = Departement::create([
+            $this->authorize('setting/manage_data/department.create');
+            $new = Department::create([
                 'name_dept' => $request->name_dept,
             ]);
             if($new){
                 return redirect()->route('dept.index')->with('msg','Data atas ('.$request->name_dept.') BERHASIL ditambahkan!');
             }
         }
-            $departement = Departement::select("*")->get();
+            $departement = Department::select("*")->get();
             // dd($data);
             return view('data.departement.index', compact('departement'));
         }
 
     public function data(Request $request){
-        $data = Departement::select('*')->orderBy("id");
+        $this->authorize('setting/manage_data/department.read');
+        $data = Department::select('*')->orderBy("id");
             return DataTables::of($data)
                     ->filter(function ($instance) use ($request) {
                         if (!empty($request->get('search'))) {
@@ -42,13 +45,14 @@ class DepartementController extends Controller
 
     public function datatables()
     {
-        $departement = Departement::select('*');
+        $departement = Department::select('*');
         return DataTables::of($departement)->make(true);
     }
    
     public function edit($id)
     {
-        $departement = Departement::findOrFail($id);
+        $this->authorize('setting/manage_data/department.update');
+        $departement = Department::findOrFail($id);
         return view('data.departement.edit', compact('departement'));
     }
     public function update(Request $request, $id)
@@ -57,15 +61,16 @@ class DepartementController extends Controller
             'name_dept' => 'required',
         ]);
 
-        $departement = Departement::findOrFail($id);
+        $departement = Department::findOrFail($id);
         $departement->update([
             'name_dept' => $request->name_dept,
         ]);
 
-        return redirect()->route('dept.index')->with('Departement,', 'Departement berhasil diperbarui.');
+        return redirect()->route('dept.index')->with('Department,', 'Department berhasil diperbarui.');
     }
     public function delete(Request $request){
-        $data = Departement::find($request->id);
+        $this->authorize('setting/manage_data/department.delete');
+        $data = Department::find($request->id);
         if($data){
             $data->delete();
             return response()->json([
@@ -80,3 +85,5 @@ class DepartementController extends Controller
         }
     }
 }
+
+
